requirements.txt
uplink
requests-http-signature
pyopenssl